python/requirements.txt
appnope==0.1.0
astroid==1.6.6
attrs==20.3.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
bleach==3.3.0
configparser==4.0.2
contextlib2==0.6.0.post1
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
enum34==1.1.10
functools32==3.2.3.post2
futures==3.3.0
importlib-metadata==2.1.1
ipaddress==1.0.23
ipykernel==4.10.1
ipython==5.10.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.5
jupyter-console==5.2.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==1.3.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==4.4.0
nose==1.3.7
notebook==5.7.10
numpy==1.16.6
packaging==20.9
pandocfilters==1.4.3
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==1.0.18
ptyprocess==0.6.0
Pygments==2.5.2
pylint==1.9.5
pyparsing==2.4.7
pyrsistent==0.16.1
pyserial==3.4
python-dateutil==2.8.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
scandir==1.10.0
scipy==1.2.3
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
terminado==0.8.3
testpath==0.4.4
tornado==5.1.1
traitlets==4.3.3
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==1.2.0
